+++ collector/processor/concurrentbatchprocessor/factory.go
@@ -18,8 +18,9 @@ const (
 
 	defaultSendBatchSize = uint32(8192)
 	defaultTimeout       = 200 * time.Millisecond
-	// default inflight bytes is 2 MiB
-	defaultMaxBytes      = 2 * 1048576
+
+	// default inflight bytes is 32 MiB
+	defaultMaxInFlightSizeMiB = 32
 
 	// defaultMetadataCardinalityLimit should be set to the number
 	// of metadata configurations the user expects to submit to
@@ -41,7 +42,7 @@ func createDefaultConfig() component.Config {
 	return &Config{
 		SendBatchSize:            defaultSendBatchSize,
 		Timeout:                  defaultTimeout,
-		MaxInFlightBytes:         defaultMaxBytes,
+		MaxInFlightSizeMiB:       defaultMaxInFlightSizeMiB,
 		MetadataCardinalityLimit: defaultMetadataCardinalityLimit,
 	}
 }