+++ collector/receiver/otelarrowreceiver/config_test.go
@@ -39,28 +39,6 @@ func TestUnmarshalConfigOnlyGRPC(t *testing.T) {
 	assert.Equal(t, defaultOnlyGRPC, cfg)
 }
 
-func TestUnmarshalOldMemoryLimitConfig(t *testing.T) {
-	cm, err := confmaptest.LoadConf(filepath.Join("testdata", "oldmemlimit.yaml"))
-	require.NoError(t, err)
-	factory := NewFactory()
-	cfg := factory.CreateDefaultConfig()
-	assert.NoError(t, component.UnmarshalConfig(cm, cfg))
-	expectCfg := factory.CreateDefaultConfig().(*Config)
-	// The number in config is <1MB, so Validate() rounds up.
-	expectCfg.Arrow.MemoryLimitMiB = 1
-	assert.NoError(t, cfg.(*Config).Validate())
-	assert.Equal(t, expectCfg, cfg)
-}
-
-func TestUnmarshalBothMemoryLimitConfig(t *testing.T) {
-	cm, err := confmaptest.LoadConf(filepath.Join("testdata", "bothmemlimit.yaml"))
-	require.NoError(t, err)
-	factory := NewFactory()
-	cfg := factory.CreateDefaultConfig()
-	assert.NoError(t, component.UnmarshalConfig(cm, cfg))
-	assert.Error(t, cfg.(*Config).Validate())
-}
-
 func TestUnmarshalConfig(t *testing.T) {
 	cm, err := confmaptest.LoadConf(filepath.Join("testdata", "config.yaml"))
 	require.NoError(t, err)