+++ pkg/otel/common/test_utils.go
@@ -20,7 +20,7 @@ package common
 import (
 	"math/rand"
 
-	"github.com/apache/arrow/go/v12/arrow"
+	"github.com/apache/arrow/go/v14/arrow"
 
 	"github.com/open-telemetry/otel-arrow/pkg/record_message"
 )