+++ pkg/otel/logs/arrow/logs.go
@@ -20,7 +20,7 @@ package arrow
 import (
 	"math"
 
-	"github.com/apache/arrow/go/v12/arrow"
+	"github.com/apache/arrow/go/v14/arrow"
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/plog"